chore: bean config 설정 부분
